Moved landing page to the welcome page, moved globals.css to style folder
