+++ src/app/Welcome/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <div>
+        <h1>Home</h1>
+    </div>
+  );
+}
