--- src/app/page.tsx
@@ -1,9 +1,0 @@
-
-
-export default function Home() {
-  return (
-    <div>
-        <h1>Home</h1>
-    </div>
-  );
-}
